third_party/WebKit/LayoutTests/fast/forms/date/input-date-validation-message-expected.txt

CONSOLE WARNING: line 20: The specified value "foo" does not conform to the required format, "yyyy-MM-dd".
Test for validationMessage IDL attribute for <input type=date>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


No message
PASS testIt("", "", "") is ""
Value missing
PASS testIt("", "", "") is "Please fill out this field."
Type mismatch
PASS testIt("foo", "", "") is ""
Range overflow
PASS testIt("1982-11-02", "", "1970-12-31") is "Value must be 12/31/1970 or earlier."
Range underflow
PASS testIt("1982-11-02", "1990-05-25", "1990-12-24") is "Value must be 05/25/1990 or later."
Step mismatch
PASS testIt("1982-11-02", "1982-01-01", "", "123") is "Please enter a valid value. The two nearest valid values are 09/04/1982 and 01/05/1983."
PASS successfullyParsed is true

TEST COMPLETE
